refactor: Dashboard 추상화, 컴포넌트 분리(MemberList)
