+++ src/pages/dashboard/components/MemberList/index.ts
@@ -1,0 +1,3 @@
+import MemberList from './MemberList';
+
+export { MemberList };
