+++ src/pages/dashboard/components/SearchAndInviteMember/index.ts
@@ -1,0 +1,3 @@
+import SearchAndInviteMember from './SearchAndInviteMember';
+
+export { SearchAndInviteMember };
