--- src/pages/dashboard/components/SearchAndInviteUser/index.ts
@@ -1,3 +1,0 @@
-import SearchAndInviteUser from './SearchAndInviteUser';
-
-export { SearchAndInviteUser };
